Refactor network formatting
- format helpers moved to a utility
- selectors refer to those utilities
